requirements.txt
python-telegram-bot==22.5
feedparser==6.0.12
requests==2.32.4
APScheduler==3.11.1
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
